refactor: apply Service Handle pattern to note service
- Define NoteServiceHandle interface as public API contract
- Move note.service.ts to note/ subdirectory
- Hide internal implementation (repositories, helpers) inside factory
- Add createMockNoteService helper for testing
- Update import paths in routes and spec files
