+++ backend/src/services/note/note.service.spec.ts
@@ -1,9 +1,9 @@
 import { describe, it, expect, beforeEach } from 'vitest';
 // NOTE: Initialize database before importing db
-import '../../tests/preload';
-import { db, notes, mentions } from '../db';
+import '../../../tests/preload';
+import { db, notes, mentions } from '../../db';
 
-import { generateId } from '../utils/id-generator';
+import { generateId } from '../../utils/id-generator';
 import { MAX_NOTE_LENGTH } from '@thread-note/shared/constants';
 import { createNoteService } from './note.service';
